Do not run focus test on Safari
